backend/requirements.txt
Flask
numpy
scikit-learn
opencv-python
joblib
tensorflow
gunicorn
Pillow
